Corregir metodos Controlador para ajustarlos a la especificacion de REST
Corregir varias pruebas en error.

+++ src/db/knex/dao/BaseKnexDao.js
@@ -31,42 +31,42 @@ export default class BaseKnexDao {
 
     }
 
-    async loadFilteredDataWithRelations(filters,start,limit,relationParams,selectQuery){
+    async loadFilteredDataWithRelations(filters, start, limit, relationParams, selectQuery) {
         let sorts = [];
 
         if (filters.sort) {
             sorts = KnexFilterParser.parseSort(filters.sort);
-        }else{
+        } else {
             sorts = 1;
         }
         let connect = KnexConnector$1.connection
-                .select(KnexConnector$1.connection.raw(selectQuery))
-                .from(this.tableName)
-                .where((builder) =>
-                    KnexFilterParser.parseFilters(builder, lodash__default['default'].omit(filters, ["sort", "start", "limit"]))
-                )
-                
-        if(relationParams){
-            if(Array.isArray(relationParams)){
+            .select(KnexConnector$1.connection.raw(selectQuery))
+            .from(this.tableName)
+            .where((builder) =>
+                KnexFilterParser.parseFilters(builder, lodash__default['default'].omit(filters, ["sort", "start", "limit"]))
+            )
+
+        if (relationParams) {
+            if (Array.isArray(relationParams)) {
                 relationParams.forEach(element => {
                     let typeInner = element.type
                     let tabletoJoinInner = element.tabletoJoin
                     let column1Inner = element.column1
                     let column2Inner = element.column2
-                    connect = connect.joinRaw( typeInner + " " + tabletoJoinInner + " ON " + column1Inner + " = " + column2Inner)
+                    connect = connect.joinRaw(typeInner + " " + tabletoJoinInner + " ON " + column1Inner + " = " + column2Inner)
                 });
-            }else{
-                    let type = relationParams.type
-                    let tabletoJoin = relationParams.tabletoJoin
-                    let column1 = relationParams.column1
-                    let column2 = relationParams.column2
-                    connect = connect.joinRaw( type + " " + tabletoJoin + " ON " + column1 + " = " + column2)
+            } else {
+                let type = relationParams.type
+                let tabletoJoin = relationParams.tabletoJoin
+                let column1 = relationParams.column1
+                let column2 = relationParams.column2
+                connect = connect.joinRaw(type + " " + tabletoJoin + " ON " + column1 + " = " + column2)
             }
         }
-      
-       return connect.orderByRaw(sorts).limit(limit).offset(start);
-           
-          
+
+        return connect.orderByRaw(sorts).limit(limit).offset(start);
+
+
     }
 
     async countFilteredData(filters) {
@@ -78,7 +78,12 @@ export default class BaseKnexDao {
     }
 
     async loadById(objectId) {
-        return KnexConnector.connection.from(this.tableName).where('id', objectId);
+        const data = await KnexConnector.connection.from(this.tableName).where('id', objectId);
+
+        if (data && data[0]) {
+            return data[0];
+        }
+        return null;
     }
 
     save(object) {
@@ -87,7 +92,11 @@ export default class BaseKnexDao {
     update(objectId, newObject) {
         return KnexConnector.connection.from(this.tableName).where("id", objectId).update(newObject).returning("*");
     }
-    delete(objectId) {
+    async delete(objectId) {
+        const existing = await this.loadById(objectId);
+        if (!existing) {
+            throw "NotFound";
+        }
         return KnexConnector.connection.from(this.tableName).where("id", objectId).delete()
     }
 }
